Extract search phrases logic from Search Form into it's own module
